pyproject.toml
[tool.poetry]
name = "mariyam-usmon"
version = "0.1.0"
description = ""
authors = ["ozoda <ozodas531@gmail.com>"]
readme = "README.md"
packages = [{include = "mariyam_usmon"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.2"
pillow = "^9.5.0"
django-mptt = "^0.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
